Self-registration - email related methods
 - Security Admins can now resend activation emails from User
 - moved email related methods from SelfRegistrationExtension to
UserExtension
 - added warning banner to Configuration.edit view for when
self-registration is enabled but either emails are not configured or the
self-registration group is not set
 - added ActivationURL field to User
 - added 'Resend Activation' button to User.edit view with conditional
visibility on whether the account is activated or not
