+++ skyve-ee/src/skyve/modules/admin/User/UserExtension.java
@@ -1,17 +1,31 @@
 package modules.admin.User;
 
 import java.util.List;
+import java.util.UUID;
 
 import org.skyve.CORE;
+import org.skyve.domain.Bean;
+import org.skyve.domain.types.DateTime;
 import org.skyve.impl.metadata.repository.AbstractRepository;
 import org.skyve.impl.metadata.user.UserImpl;
 import org.skyve.impl.persistence.hibernate.AbstractHibernatePersistence;
 import org.skyve.impl.util.SQLMetaDataUtil;
 import org.skyve.impl.util.UtilImpl;
 import org.skyve.metadata.customer.Customer;
+import org.skyve.metadata.model.document.Document;
+import org.skyve.metadata.model.document.Relation;
 import org.skyve.metadata.module.Module;
 import org.skyve.metadata.user.Role;
+import org.skyve.persistence.Persistence;
+import org.skyve.util.BeanVisitor;
+import org.skyve.util.Binder;
+import org.skyve.util.Util;
 
+import modules.admin.Communication.CommunicationUtil;
+import modules.admin.Communication.CommunicationUtil.ResponseMode;
+import modules.admin.domain.Contact;
+import modules.admin.domain.SelfRegistration;
+import modules.admin.domain.SelfRegistrationActivation;
 import modules.admin.domain.User;
 import modules.admin.domain.UserRole;
 
@@ -20,6 +34,19 @@ public class UserExtension extends User {
 
 	private boolean determinedRoles = false;
 
+	public static final String SELF_REGISTRATION_COMMUNICATION = "SYSTEM Self Registration";
+	public static final String SELF_REGISTRATION_SUBJECT = "Activate your account";
+	public static final String SELF_REGISTRATION_HEADING = "Welcome!";
+	public static final String SELF_REGISTRATION_BODY = String.format("<p>Hi <b>{%1$s}</b>,</p><br/>"
+			+ "<p>Thank you for registering.</p>"
+			+ "<p>To complete your account setup, please click the activation link below.</p>"
+			+ "<p><a href=\"{%2$s}\">{%2$s}</a></p>"
+			+ "<p>If you have any questions about your new account, contact us at <a href=\"mailto:%3$s\">%3$s</a>.</p>",
+			Binder.createCompoundBinding(User.contactPropertyName, Contact.namePropertyName),
+			activateUrlPropertyName,
+			Util.getSupportEmailAddress());
+
+	
 	@Override
 	public List<UserRole> getAssignedRoles() {
 		List<UserRole> assignedRoles = super.getAssignedRoles();
@@ -75,5 +102,102 @@ public class UserExtension extends User {
 		
 		return metaDataUser;
 	}
+	
+	/**
+	 * Generates the activation code and link for this new user and upserts them
+	 * into the datastore.
+	 */
+	public void generateActivationDetailsAndSave(final Persistence persistence) {
+		// Set activation details
+		this.setActivated(Boolean.FALSE);
+		this.setActivationCode(UUID.randomUUID().toString());
+		this.setActivationCodeCreationDateTime(new DateTime());
 
+		this.setBizUserId(getBizId());
+		
+		// Save and set the user
+		this.upsertUser(persistence, this);
+		
+		// Generate link used for activation
+		generateActivationLink();
+	}
+	
+	/**
+	 * Generates the activation link for the email to send to the new user with the activation code.
+	 */
+	public void generateActivationLink() {
+		StringBuilder urlBuilder = new StringBuilder();
+		urlBuilder.append(Util.getDocumentUrl(SelfRegistrationActivation.MODULE_NAME, SelfRegistrationActivation.DOCUMENT_NAME))
+				.append("&code=")
+				.append(this.getActivationCode());
+		this.setActivateUrl(urlBuilder.toString());
+	}
+	
+	/**
+	 * Sends the activation email to the user who registered.
+	 * 
+	 * @throws Exception
+	 */
+	public void sendUserRegistrationEmail() throws Exception {
+		Util.LOGGER.info("Sending registration email to " + this.getContact().getEmail1());
+		CommunicationUtil.sendFailSafeSystemCommunication(SELF_REGISTRATION_COMMUNICATION,
+				this.getContact().getEmail1(),
+				null,
+				SELF_REGISTRATION_SUBJECT,
+				SELF_REGISTRATION_BODY,
+				ResponseMode.EXPLICIT,
+				null,
+				this);
+	}
+
+	/**
+	 * Upsert the user into the database.
+	 * <br />
+	 * NOTE: We upsert the user rather than regularly calling {@link Persistence#save(org.skyve.domain.PersistentBean)} because
+	 * calling save will automatically set the bizUserId to be the current logged in user.
+	 * <br />
+	 * When registering, we want the User we are just about to create to own the documents.
+	 * 
+	 * @param persistence skyve persistence to save the bean
+	 * @param bean the user bean to register
+	 * @return the saved user bean
+	 */
+	private UserExtension upsertUser(Persistence persistence, UserExtension bean) {
+		persistence.begin();
+
+		// update bizuser id on User and related objects
+		org.skyve.metadata.user.User u = persistence.getUser();
+		Customer c = u.getCustomer();
+		Module am = c.getModule(bean.getBizModule());
+		Document ad = am.getDocument(c, bean.getBizDocument());
+		new UpdateBizUserVisitor(bean.getBizId()).visit(ad, bean, c);
+
+		// upsert Contact, User and Roles
+		persistence.upsertBeanTuple(bean.getContact());
+		persistence.upsertBeanTuple(bean);
+		persistence.upsertCollectionTuples(bean, User.groupsPropertyName);
+
+		persistence.commit(false);
+		return bean;
+	}
+	
+	/**
+	 * Visitor to update all beans related to a document to have a given owning bizUserId
+	 */
+	private class UpdateBizUserVisitor extends BeanVisitor {
+
+		private String bizUserId;
+
+		public UpdateBizUserVisitor(String bizUserId) {
+			super(false, false, false);
+			this.bizUserId = bizUserId;
+		}
+
+		@Override
+		protected boolean accept(String binding, Document document, Document owningDocument, Relation owningRelation, Bean bean) {
+			bean.setBizUserId(bizUserId);
+			return true;
+		}
+	}
+	
 }
