+++ skyve-ee/src/skyve/modules/admin/admin.xml
@@ -332,6 +332,7 @@
 					<action name="New" />
 					<action name="Next" />
 					<action name="Back" />
+					<action name="ResendActivation"/>
 				</document>
 				<document name="UserCandidateContact" permission="_____" />
 				<document name="UserList" permission="_____">
